api/requirements.txt
Django==5.2.5
djangorestframework==3.15.2
django-cors-headers==4.6.0
drf-spectacular==0.27.2
openai>=1.55.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0

